Create an image to deploy the notebook as well
This medium post
https://medium.com/@patternrecognizer/how-to-add-a-try-in-play-with-docker-button-to-your-github-project-41cb65721e94
claimed that "Try in PWD" works with the `build` tag as well - e.g.

```
services:
  redis-commander:
    build: https://github.com/joeferner/redis-commander.git
```

and cites https://stackoverflow.com/questions/34120504/how-can-i-make-docker-compose-build-an-image-from-a-remote-git-repository/34120804#34120804

However, this only works for `docker-compose`. And as we can see from the pwd code https://github.com/play-with-docker/play-with-docker/blob/3bfb09802a6b41ad98adfee79535f813323b90a9/pwd/session.go#L183, it uses `docker stack deploy` to deploy.

And `docker stack deploy` does not support the `build` tag https://stackoverflow.com/questions/41472219/docker-usage-in-compose-swarm-mode - it only works with published images.

So this change:
- adds instructions to build the notebook using the `nomkl` server image as the base
- changes the `docker-compose-small` file to use the pre-built image instead of rebuilding