Simple fix/hack to support a limited set of indices
DocumentDB is the AWS RDS that is similar to MongoDB. Although it supports most
queries, it has limited support for indices.

To solve this, we introduce a new environment variable SIMPLE_INDICES.
If SIMPLE_INDICES is set, we remove all HASHED and GEOSPHERE indices from the database driver.
We currently do this by editing the file directly, which is a giant hack but works.

If we want to support DocumentDB in the long term, we should move this check to
the source code directly.

This fixes
https://github.com/e-mission/e-mission-docs/issues/597

Testing done:

With this patch to the `docker-compose.yml`

```
diff --git a/examples/em-server-multi-tier-cronjob/docker-compose.yml b/examples/em-server-multi-tier-cronjob/docker-compose.yml
index 75d3f20..9671f61 100644
--- a/examples/em-server-multi-tier-cronjob/docker-compose.yml
+++ b/examples/em-server-multi-tier-cronjob/docker-compose.yml
@@ -8,6 +8,7 @@ services:
       - DB_HOST=db
       - WEB_SERVER_HOST=0.0.0.0
       - SERVER_BRANCH=ceo_ebike_project
+      - SIMPLE_INDICES=1
     deploy:
       replicas: 1
       restart_policy:
@@ -27,6 +28,8 @@ services:
       - DB_HOST=db
       - WEB_SERVER_HOST=0.0.0.0
       - SERVER_BRANCH=ceo_ebike_project
+      - SIMPLE_INDICES=1
+
     deploy:
       replicas: 1
       restart_policy:
```

the webserver indices are replaced

```
web-server_1       | Found configuration, overriding...
web-server_1       | Live reload disabled,
web-server_1       | Replacing database indices for compatibility with DocumentDB
```

the analysis indices are replaced

```
analysis-server_1  | Found configuration, overriding...
analysis-server_1  | Live reload disabled,
analysis-server_1  | Replacing database indices for compatibility with DocumentDB
```